+++ aws/eks-helm/aws-load-balancer-controller/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 ## [0.1.41]() (2025-01-08)
 # Features
-* Add new policy document with new permissions for ALB controller version >= 2.9. Ref: https://github.com/kubernetes-sigs/aws-load-balancer-controller/releases/tag/v2.9.0
+* Add new policy documents for chart versions and auto get policy version based on chart_version.
 
 ## [0.1.20]() (2024-12-24)
 ### Features